MVVM pattern included
I modify all the necessary files to follow the MVVM architecture pattern, including creating a LoginViewModel file and adding it to the Viewmodels folder.
